+++ programs/drift/src/state/state.rs
@@ -35,7 +35,8 @@ pub struct State {
     pub exchange_status: u8,
     pub liquidation_duration: u8,
     pub initial_pct_to_liquidate: u16,
-    pub padding: [u8; 14],
+    pub max_number_of_sub_accounts: u16,
+    pub padding: [u8; 12],
 }
 
 #[derive(BitFlags, Clone, Copy, PartialEq, Debug, Eq)]
@@ -72,6 +73,12 @@ impl State {
         Ok(self
             .get_exchange_status()?
             .contains(ExchangeStatus::FundingPaused))
+    }
+
+    pub fn max_number_of_sub_accounts(&self) -> u64 {
+        (self.max_number_of_sub_accounts as u64)
+            .saturating_mul(100)
+            .max(25000)
     }
 }